[Move] #3  viewModel 파일 이동
