+++ app/src/main/java/com/project/drinkly_admin/viewModel/UserViewModel.kt
@@ -21,8 +21,6 @@ class UserViewModel : ViewModel() {
 
     var storeList = MutableLiveData<MutableList<StoreListResponse>>()
 
-    var storeDetailInfo: MutableLiveData<StoreDetailResponse> = MutableLiveData()
-
     fun getOwnerName(activity: MainActivity) {
         val apiClient = ApiClient(activity)
         val tokenManager = TokenManager(activity)
@@ -102,41 +100,6 @@ class UserViewModel : ViewModel() {
                 }
 
                 override fun onFailure(call: Call<BaseResponse<List<StoreListResponse>>>, t: Throwable) {
-                    // 통신 실패
-                    Log.d("DrinklyViewModel", "onFailure 에러: " + t.message.toString())
-                }
-            })
-    }
-
-    fun getStoreDetail(activity: MainActivity, storeId: Int) {
-        val apiClient = ApiClient(activity)
-        val tokenManager = TokenManager(activity)
-
-        apiClient.apiService.getStoreDetailInfo(tokenManager.getAccessToken().toString(), storeId)
-            .enqueue(object :
-                Callback<BaseResponse<StoreDetailResponse>> {
-                override fun onResponse(
-                    call: Call<BaseResponse<StoreDetailResponse>>,
-                    response: Response<BaseResponse<StoreDetailResponse>>
-                ) {
-                    Log.d("DrinklyViewModel", "onResponse 성공: " + response.body().toString())
-                    if (response.isSuccessful) {
-                        // 정상적으로 통신이 성공된 경우
-                        val result: BaseResponse<StoreDetailResponse>? = response.body()
-                        Log.d("DrinklyViewModel", "onResponse 성공: " + result?.toString())
-
-                        storeDetailInfo.value = result?.payload!!
-                    } else {
-                        // 통신이 실패한 경우(응답코드 3xx, 4xx 등)
-                        var result: BaseResponse<StoreDetailResponse>? = response.body()
-                        Log.d("DrinklyViewModel", "onResponse 실패: " + response.body())
-                        val errorBody = response.errorBody()?.string() // 에러 응답 데이터를 문자열로 얻음
-                        Log.d("DrinklyViewModel", "Error Response: $errorBody")
-
-                    }
-                }
-
-                override fun onFailure(call: Call<BaseResponse<StoreDetailResponse>>, t: Throwable) {
                     // 통신 실패
                     Log.d("DrinklyViewModel", "onFailure 에러: " + t.message.toString())
                 }
